check for existence in bootstrap

+++ bootstrap.sh
@@ -24,7 +24,9 @@ export GOPATH="${PWD}/.bootstrap:${PWD}"
 if [ -z "${PLZ_NO_GO_BOOTSTRAP+bootstrap}" ]; then
     noticen "Installing Go dependencies..."
     mkdir -p "${PWD}/.bootstrap/src/github.com/thought-machine"
-    ln -s "$PWD" "${PWD}/.bootstrap/src/github.com/thought-machine/please"
+    if [ ! -e "${PWD}/.bootstrap/src/github.com/thought-machine/please" ]; then
+        ln -s "$PWD" "${PWD}/.bootstrap/src/github.com/thought-machine/please"
+    fi
     go_get golang.org/x/crypto/ssh/terminal
     go_get golang.org/x/sync/errgroup
     go_get golang.org/x/tools/cover
